Delete provisioned volumes on binding conflict.
When we provision a volume for a claim and at the same time the claim is bound
to different volume we should delete the volume.

For example:
1. user submits a PVC
2. synPVC() does not find a suitable PV for it and starts provisioning
3. admin adds some new PVs manually
4. syncPVC() finds a suitable PV (pv2) for the PVC and binds it
5. provisioner finishes and submits a provisioned pv1 to etcd. Lets's say this
   PV has:
   pv1.ClaimRef.UID set to the pvc.UID
   pv1.Annotations[annBoundByController] set
6. Now we have pv1 that points to the PVC and the PVC points to pv2.
   syncPV() ends up here:
   // This is part of the normal operation of the controller;
   // the controller tried to use this volume for a claim but the claim
   // was fulfilled by another volume.
   // We did this; fix it.
   pv.Spec.ClaimPtr = nil
   ...
   pv.Status.Phase = Available

The provisioned volume becomes Available. This is wrong, we should delete it
instead.

Now, if we create the provisioned PV with no annBoundByController, we reach
similar part of syncPV:
// The PV was created with this pointer, but the claim is
// bound to another volume already. This PV is considered to
// be 'Available', in the sense that it is not bound, even
// though the set of PVs it can bind to is restricted to a
// specific PVC.
pv.Status.Phase = Available

Again, this is wrong. The PV then waits until the PVC is deleted and then it
gets deleted (based on its recycle policy). IMO we should delete the volume
immediately.